requirements.txt
moviepy==1.0.3
wheel
tqdm
pytorch_lightning
black
ruff
hydra-core
jaxtyping
beartype
wandb
einops
colorama
scikit-image
colorspacious
matplotlib
imageio
timm
dacite
lpips
e3nn
plyfile
tabulate
svg.py
opencv-python==4.6.0.66
sk-video
addict
spconv-cu118
